vresutils.graph: Get rid of buggy NodeListBFS and simplify polygon_subgraph
NodeListBFS's node relabeling causes buggy behaviour and is not at all
performant in what it does. Instead, we extend the BreadthFirstLevels to
cope with a set of nodes as the root set and can drastically simplify
polygon_subgraph that way.

Also using G.subgraph instead of copying and then removing nodes of a
subgraph is preferable.

Conflicts:
	graph.py